chat_with_human/eval/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
google-adk[eval]
python-dotenv